Update token handling to support larger counts
Changed InputTokens and OutputTokens in Session.cs from int to long. Updated corresponding migration files to reflect these changes in the database schema. Added logic in ChatService.cs to track token consumption during streaming responses. Created new migration files to ensure the database is aligned with the updated model definitions. These changes enhance the application's ability to handle larger token values and improve tracking of token usage.

+++ RR.AI-Chat/RR.AI-Chat.Repository/Migrations/20250805221917_Initial.cs
@@ -42,8 +42,8 @@ namespace RR.AI_Chat.Repository.Migrations
                     Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
                     Name = table.Column<string>(type: "nvarchar(100)", maxLength: 100, nullable: true),
                     Conversations = table.Column<string>(type: "nvarchar(max)", nullable: true),
-                    InputTokens = table.Column<int>(type: "int", nullable: false),
-                    OutputTokens = table.Column<int>(type: "int", nullable: false),
+                    InputTokens = table.Column<long>(type: "bigint", nullable: false),
+                    OutputTokens = table.Column<long>(type: "bigint", nullable: false),
                     DateModified = table.Column<DateTime>(type: "datetime2", nullable: false),
                     DateCreated = table.Column<DateTime>(type: "datetime2", nullable: false),
                     DateDeactivated = table.Column<DateTime>(type: "datetime2", nullable: true)
